refactor(build): use gulp with typescript & split scripts
