+++ scripts/gulp/tasks/config.ts
@@ -1,0 +1,2 @@
+export const BUILD_DIR = './.ng_build';
+export const LIB_DIR = './src/.lib';
